refactor: major data fetching overhaul
started as a light fix for classic details request
and obtaining less information from logs, became:
- removed major logs
- getting evidence values from subgraph
- unified request timelines

major bugs are to be expected.
during this development a regression was introduced
into the subgraph that the frontend was using,
urging me to push this PR so that at least some
interactions remain possible.
pending tasks:
- review appealing rendering and logic
- issues in subgraph are preventing fetching some
information related to arbitrator in light curate
- change subgraph endpoint to kleros owned